services/vector-db-service/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
python-dotenv 